+++ test/test_language_detection.ts
@@ -1,7 +1,11 @@
 "use strict";
 
 import { expect } from "chai";
+import { getLogger } from "log4js";
 import { LanguageDetection } from "../src/index";
+
+const logger = getLogger();
+logger.level = "off";
 
 describe("language detection", () => {
 
@@ -16,27 +20,33 @@ describe("language detection", () => {
     it("should detect English", (done) => {
 
       languageDetection.language(textEnglish).then((body) => {
-        expect(body).to.have.property("langCode").to.be.equal("en");
-        expect(body).to.have.property("confidence").to.be.equal(1);
-        expect(body).to.have.property("langStr").to.be.equal("English");
+        logger.debug("english:", JSON.stringify(body, null, "  "));
+        expect(body).to.have.property("detections").to.be.an("array").have.lengthOf(1);
+        expect(body.detections[0]).to.have.property("langCode").to.be.equal("en");
+        expect(body.detections[0]).to.have.property("confidence").to.be.equal(1);
+        expect(body.detections[0]).to.have.property("langStr").to.be.equal("English");
       }).then(done, done);
     });
 
     it("should detect German", (done) => {
 
       languageDetection.language(textDeutsch).then((body) => {
-        expect(body).to.have.property("langCode").to.be.equal("de");
-        expect(body).to.have.property("confidence").to.be.equal(1);
-        expect(body).to.have.property("langStr").to.be.equal("German");
+        logger.debug("german:", JSON.stringify(body, null, "  "));
+        expect(body).to.have.property("detections").to.be.an("array").have.lengthOf(1);
+        expect(body.detections[0]).to.have.property("langCode").to.be.equal("de");
+        expect(body.detections[0]).to.have.property("confidence").to.be.equal(1);
+        expect(body.detections[0]).to.have.property("langStr").to.be.equal("German");
       }).then(done, done);
     });
 
     it("should detect Spanish", (done) => {
 
       languageDetection.language(textEspañol).then((body) => {
-        expect(body).to.have.property("langCode").to.be.equal("es");
-        expect(body).to.have.property("confidence").to.be.equal(1);
-        expect(body).to.have.property("langStr").to.be.equal("Spanish");
+        logger.debug("spanish:", JSON.stringify(body, null, "  "));
+        expect(body).to.have.property("detections").to.be.an("array").have.lengthOf(1);
+        expect(body.detections[0]).to.have.property("langCode").to.be.equal("es");
+        expect(body.detections[0]).to.have.property("confidence").to.be.equal(1);
+        expect(body.detections[0]).to.have.property("langStr").to.be.equal("Spanish");
       }).then(done, done);
     });
 
@@ -46,7 +56,8 @@ describe("language detection", () => {
 
     it("should provide version", (done) => {
       languageDetection.version().then((body) => {
-        expect(body).to.have.property("version").to.be.equal("1.0-SNAPSHOT");
+        logger.debug("version:", JSON.stringify(body, null, "  "));
+        expect(body).to.have.property("version").to.be.equal("1.0.4");
       }).then(done, done);
     });
   });